ai-service/requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
Pillow==11.0.0
